Remove DiskStore's dependency on BlockManager.
